+++ includes/class-plugin-name-i18n.php
@@ -1,5 +1,5 @@
 <?php
-
+namespace PluginName;
 /**
  * Define the internationalization functionality
  *
@@ -35,7 +35,7 @@ class Plugin_Name_i18n {
 	public function load_plugin_textdomain() {
 
 		load_plugin_textdomain(
-			'plugin-name',
+			PLUGIN_NAME_TEXT_DOMAIN,
 			false,
 			dirname( dirname( plugin_basename( __FILE__ ) ) ) . '/languages/'
 		);